merge 175 into 112

+++ blender_operations.py
@@ -4,10 +4,9 @@ from os import path
 import cv2
 import io
 import tempfile
-import time
 from .image_processing import PlaneItem, EditPicture, SaveImage
 from .file_conversion import blend_opener, fbx_opener
-from .DepthByColor import GenerateEdges, NormaliseData, GenerateShapeEdges, GetDistanceBetweenPoints, ColorCheck, ResetNormals, CountVerticesFromCamera, visiblePoints
+from .DepthByColor import GenerateEdges, NormaliseData, GenerateShapeEdges, GetDistanceBetweenPoints, ColorCheck, ResetNormals, CountVerticesFromCamera, MultipleImagePath
 
 #saveObj
 #Description
@@ -325,7 +324,6 @@ def DrawMeshToScreen(MeshStructure, self, CollectionName = "Sketch_to_Mesh_Colle
         # add object to scene collection
         collection.objects.link(new_object)
         ResetNormals(CollectionName)
-        CountVerticesFromCamera()
 
 
 #DrawMesh
@@ -338,31 +336,42 @@ def DrawMeshToScreen(MeshStructure, self, CollectionName = "Sketch_to_Mesh_Colle
 #PlaneArray: This is the list of planes used for mesh creation
 #isComplex: thhis tells us if we are making a complex image or not
 
-def DrawMesh(ColorWeAreLookingFor, PolyCount, self, PlaneArray:list[PlaneItem], isComplex):
+def DrawMesh(ColorWeAreLookingFor, PolyCount, self, PlaneArray: list[PlaneItem], isComplex):
+    counter = 0
+    MeshStructure = {}
     for plane in PlaneArray:
-        if isComplex == True: #only happens when complex is called
+        if isComplex == True:
+            # Processing for complex images
             ImageDictionary, Imagedata, Image = GetlistOfPixels(PolyCount, ColorWeAreLookingFor, plane)
             VertList = NormaliseVertList(ImageDictionary)
-
-            if VertList == False:  return False #ends the function before any extra work is done
+            if VertList is False:
+                self.report({'ERROR'}, "Invalid Image")
+                continue  # Skip processing if normalization fails
             MeshStructure = GenerateEdges(VertList, "BlenderPoints")
-
-            DrawMeshToScreen(MeshStructure, self, "Sketch_To_Mesh_Collection") # this is the outline of the mesh
-            PixelLineDictionary= GetLineOfPixels(ImageDictionary[0], Imagedata, Image, ColorWeAreLookingFor)
+            DrawMeshToScreen(MeshStructure, self, "Sketch_To_Mesh_Collection")
+            PixelLineDictionary = GetLineOfPixels(ImageDictionary[0], Imagedata, Image, ColorWeAreLookingFor)
             MeshStructureLibrary = DefineDictioniesForColorsLines(PixelLineDictionary)
             DrawMeshToScreen(MeshStructureLibrary[0], self, "Sketch_To_Mesh_Collection")
-        else: 
-            MeshStructure = GenerateShapeEdges(PolyCount, plane, ColorWeAreLookingFor) #only called when not complex is called
-            DrawMeshToScreen(MeshStructure, self) #draws all the meshes to screen
-           
-# TODO: return something that is not 0. case handling and error handling, as well as completed and noncompleted states.
+            break #temp break    
+        else:
+            # Processing for simple images
+            if counter < 1: MeshStructure, EdgeData = GenerateShapeEdges(PolyCount, plane, ColorWeAreLookingFor)
+            else: MeshStructure = MultipleImagePath(MeshStructure, EdgeData, PolyCount, plane, ColorWeAreLookingFor)
+            for ImageValues in range(3):
+                CurrPath = os.path.abspath("ImageFolder\\" + "View" + str(ImageValues) + plane.PlaneFilepath[plane.PlaneFilepath.rfind("."): ] )
+                if path.exists(CurrPath): os.remove(CurrPath) # if we find that file we will delete it
+        counter += 1
+        break #temp break
+        
+    if MeshStructure: DrawMeshToScreen(MeshStructure, self)
+    else: self.report({'ERROR'}, "Invalid Image")
+
 def encode_file(file_path):
     
    with open(file_path, "rb") as file:
         blend_file_contents = io.BytesIO(file.read())
         return blend_file_contents
 
-# TODO: return something that is not 0. case handling and error handling, as well as completed and noncompleted states.
 def decode_file(data, file_extension):
     #Apparently the data doesn't need to be decoded so we will handle the different
     #file extensions handled here instead of outside the file_conversion.py file
@@ -374,12 +383,9 @@ def decode_file(data, file_extension):
 
     #Deal with the separate file extensions
     match file_extension :
-        case ".blend":
-            blend_opener(temp_file.name)
-        case ".fbx":
-            fbx_opener(temp_file.name)
-        case _: #defualt case # if there is an image file
-            bpy.ops.import_image.to_plane(files=[{"name":temp_file.name, "name":temp_file.name}], directory="", relative=False)
+        case ".blend": blend_opener(temp_file.name)
+        case ".fbx":fbx_opener(temp_file.name)
+        case _:  bpy.ops.import_image.to_plane(files=[{"name":temp_file.name, "name":temp_file.name}], directory="", relative=False)#defualt case # if there is an image file
 
     #remove the temp file
     os.unlink(temp_file.name)
@@ -409,17 +415,15 @@ def SearchForClosestPoint(PointArray, startingPoint ):
 
 
 def ResetImage(GlobalPlaneDataArray:list[PlaneItem]):
-    Itervalue = 0
 
     for plane_data in GlobalPlaneDataArray :
         #Finds the Images Saved
-        ImageFilePaths = []
-        for ImageValues in range(3):
-            ImageFilePaths.append(os.path.abspath("ImageFolder\\" + "View" + str(ImageValues) + plane_data.PlaneFilepath[plane_data.PlaneFilepath.rfind("."): ] ))
-            if path.exists(ImageFilePaths[ImageValues]): os.remove(ImageFilePaths[ImageValues]) # if we find that file we will delete it
-            
-        bpy.data.objects[plane_data.ImagePlaneName].select_set(True)# selects the image plane in the array 
-        bpy.ops.object.delete(use_global=False, confirm=False)#deletes the image plane in the array 
-        Itervalue = Itervalue + 1 #increases the itervalue to reach the next View string
+        if plane_data.ImagePlaneName != "":
+
+            ImageFilePaths = []
+            for ImageValues in range(3):
+                ImageFilePaths.append(os.path.abspath("ImageFolder\\" + "View" + str(ImageValues) + plane_data.PlaneFilepath[plane_data.PlaneFilepath.rfind("."): ] ))
+                if path.exists(ImageFilePaths[ImageValues]): os.remove(ImageFilePaths[ImageValues]) # if we find that file we will delete it
     # clears the PlaneData Array
     GlobalPlaneDataArray.clear() 
+
